refactor: split cli frontdoor responsibilities
Move query-first routing, machine-entry handling, and standalone root commands into focused modules while keeping polylogue.cli.click_app as the stable public entry surface.
